The redundant time infomation is not displayed: In the tutor hint in elec6b:      "the angle between the displacement of e between T0 and T1 and the electric force on e       by the unspecified between T0 and T1" The hint is now displayed with time only at the last:      (when two are in the same time interval):      "the angle between the displacement of e, and the electric force on e between T0 and T1"
Issue: The redundant time infomation should be allowed in the input. In the current modification, all reduadant time information was removed from expansion. It was because the modification was done outside (conjoin .. ) process. The ideal version would be using "allowed" tag instead of removal of redundant time info. It should be done by modifying (conjoin ..) process.