Working on using match ontology as a parser.  Outstanding problems:
Recursion of bindings as we recurse through members of ontology
is broken.  We currently use a single binding list so any time
there are multiple operators using the same variable name, we
have problems.  Two solutions:  munge the variable names (like is done
in problem generation), or make the bindings a list of lists.

The other problem is that strings in the ontology need to be
expanded into a list of words.