Simplify active download reference counting
Count active downloads in the watchDownloads callback rather than trying to track them when calling download/cancel/etc. This keeps the logic in one place and allows us to survive a situation where the thread with the file locker dies, but the thread with the download transfers keeps going, and we reload the app to resume.